src/content_composer/__init__.py
from dotenv import load_dotenv

from .langgraph_workflow import execute_workflow
from .recipe_parser import parse_recipe

load_dotenv()

# Import core functions to ensure they are registered
from . import core_functions

__all__ = [
    "parse_recipe",
    "load_recipe",
    "execute_workflow",
]
    